+++ src/clj/rems/css/styles.clj
@@ -20,42 +20,24 @@
             [mount.core :as mount]
             [rems.config :refer [env]]
             [rems.context :as context]
+            [rems.css.style-utils :refer [get-theme-attribute get-logo-image get-navbar-logo get-logo-name-sm]]
             [ring.util.response :as response]))
-
-(defn get-theme-attribute
-  "Fetch the attribute value from the current theme with fallbacks.
-
-  Keywords denote attribute lookups while strings are interpreted as fallback constant value."
-  [& attr-names]
-  (when (seq attr-names)
-    (let [attr-name (first attr-names)
-          attr-value (if (keyword? attr-name)
-                       (get (:theme env) attr-name)
-                       attr-name)]
-      (or attr-value (recur (rest attr-names))))))
 
 (def content-width (u/px 1200))
 (def logo-height-menu (u/px 40))
 (def logo-height (u/px 150))
 (def menu-height 56)
 
-(defn resolve-image [path]
-  (when path
-    (let [url (if (str/starts-with? path "http")
-                path
-                (str (get-theme-attribute :img-path "../../img/") path))]
-      (str "url(\"" url "\")"))))
+;; Fonts of the app 
 
-(defn get-logo-image [lang]
-  (resolve-image (get-theme-attribute (keyword (str "logo-name-" (name lang))) :logo-name)))
-
-(defn get-logo-name-sm [lang]
-  (resolve-image (get-theme-attribute (keyword (str "logo-name-" (name lang) "-sm")) :logo-name-sm)))
-
-(defn get-navbar-logo [lang]
-  (resolve-image (get-theme-attribute (keyword (str "navbar-logo-name-" (name lang))) :navbar-logo-name)))
-
-(defn- generate-at-font-faces []
+(defn generate-at-font-faces
+  "The theme :font-family settings will override these fonts is set.
+  
+  Reason for this function to be included into our screen.css
+   - because themes like Findata can use them since our theme.edn doesn't offer a possibility to add custom fonts
+   - all the fonts need to be 'built in' to REMS, even if they're not active
+   - bundling Robot Slab and Lato in REMS is a bit of a hack, made for Findata & THL to get the sort of styles they want"
+  []
   (list
    (stylesheet/at-font-face {:font-family "'Lato'"
                              :src "url('/font/Lato-Light.eot')"}
@@ -73,7 +55,7 @@
                              :font-weight 700
                              :font-style "normal"})
    (stylesheet/at-font-face {:font-family "'Roboto Slab'"
-                             :src "url('/font/Roboto.woff2') format('woff2')"
+                             :src "url('/font/Roboto-Slab.woff2') format('woff2')"
                              :font-weight 400
                              :font-style "normal"})))